Make DDP tests backend agnostic
DDP tests were tightly coupled to NCCL, limiting portability.

- Separate DistributedDataParallelTest from test_c10d_nccl.py into
  new backend-agnostic test_ddp.py
- Refactor test_ddp_hooks.py to be backend agnostic
